+++ home/www/dashboard/public/index.php
@@ -71,7 +71,14 @@ $postgresDatabases = $postgres->query('SELECT datname AS name FROM pg_database W
                                 ?>
 
                                 <?= $site ?>
-                                <span class="badge badge-primary">Link</span>
+                                <span
+                                    class="badge badge-primary"
+                                    data-toggle="tooltip"
+                                    data-placement="right"
+                                    data-html="true"
+                                    title="Linked project container(s) required. Run <code>ops link</code> In project directory.">
+                                        link
+                                </span>
 
                                 <?php
                             } else {
@@ -301,5 +308,11 @@ $postgresDatabases = $postgres->query('SELECT datname AS name FROM pg_database W
         <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
         <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+
+        <script>
+            $(function () {
+                $('[data-toggle="tooltip"]').tooltip();
+            });
+        </script>
     </body>
 </html>