+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- **dashboard** Tooltip for 'live' project badge
+- `ops system refresh-certs` command
+- `ops system refresh-config` command
+- `ops system refresh-services` command
+
 ## [0.5.0] - 2018-05-24
 
 ### Added